+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [0.0.5] 02-Apr-2019
+ - Add arabic Language
+ - add order number word for arabic
+## [0.0.5] 02-Apr-2019
  - Add Indonesian Language
  - Add English_US Language
 ## [0.0.1] 01-Apr-2019